--- a/unit_tests/util/query_heap.cpp
+++ b/unit_tests/util/query_heap.cpp
@@ -22,9 +22,8 @@ struct TestData
 using TestNodeID = NodeID;
 using TestKey = int;
 using TestWeight = int;
-using storage_types = boost::mpl::list<ArrayStorage<TestNodeID, TestKey>,
-                                       TwoLevelStorage<TestNodeID, TestKey>,
-                                       UnorderedMapStorage<TestNodeID, TestKey>>;
+using storage_types =
+    boost::mpl::list<ArrayStorage<TestNodeID, TestKey>, UnorderedMapStorage<TestNodeID, TestKey>>;
 
 template <unsigned NUM_ELEM> struct RandomDataFixture
 {